pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opensearch-docling-graphrag"
version = "0.1.0"
description = "Fully local RAG: OpenSearch + Neo4j + Ollama + Docling"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.0.0",
    "opensearch-py>=2.4.0",
    "neo4j>=5.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "streamlit>=1.30.0",
    "pyvis>=0.3.2",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["opensearch_graphrag*", "api*", "ui*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"ui/*" = ["E402"]
"scripts/*" = ["E402"]
"run_api.py" = ["E402"]
